Sesion 1/requirements.txt
langchain
langchain-aws
boto3
python-dotenv
